ExplorerRpc::GetTransaction
