+++ src/pocketdb/web/PocketTransactionRpc.cpp
@@ -44,24 +44,6 @@ namespace PocketWeb::PocketWebRpc
         return estimatesmartfee(request);
     }
 
-    UniValue GetTransaction(const JSONRPCRequest& request)
-    {
-        if (request.fHelp)
-            throw runtime_error(
-                "getrawtransaction\n"
-                "\nGet transaction data.\n"
-            );
-
-        RPCTypeCheck(request.params, {UniValue::VSTR});
-        string txid = request.params[0].get_str();
-
-        UniValue txs = request.DbConnection()->ExplorerRepoInst->GetTransactions({ txid }, 1, 2);
-        if (txs.empty())
-            return UniValue(UniValue::VOBJ);
-
-        return txs[0];
-    }
-
     CMutableTransaction ConstructPocketnetTransaction(const UniValue& inputs_in, const CTxOut& dataOut,
         const UniValue& outputs_in, const UniValue& locktime, const UniValue& rbf)
     {
