+++ src/pocketdb/web/PocketTransactionRpc.h
@@ -26,7 +26,6 @@ namespace PocketWeb::PocketWebRpc
 {
     UniValue _accept_transaction(const CTransactionRef& tx, const PTransactionRef& ptx);
     UniValue AddTransaction(const JSONRPCRequest& request);
-    UniValue GetTransaction(const JSONRPCRequest& request);
     UniValue EstimateSmartFee(const JSONRPCRequest& request);
     UniValue GenerateTransaction(const JSONRPCRequest& request);
     UniValue GenerateAddress(const JSONRPCRequest& request);
